ANDROID-1203: Frambo: at times supplying incorrect wifi pw returns success in app
Changed DeviceWifiSetup.start to subscribe to DeviceModel.getDeviceSyncPostUpdateObservable (instead of getUpdateObservable) and filter on the specific wifi attributes we care about. This ensures we only emit from DeviceWifiSetup.observeWifiSetup when an actual change to the state attribute comes in, as opposed to some other device change. Also we no longer emit the current setup state when subscribed. This is less likely to confuse app side logic especially for devices (Frambo) that are slow to send up attribute updates during the setup flow.