ANDROID-1240: Add support for persistent device tags
Changed AferoClient.putDeviceTag() to take a DeviceTag object as a param so we can properly copy over all the relevant traits for the PUT.
Rewrote DeviceTagAdapter to take advantage of SortedList & SortedListAdapterCallback.
Changed DeviceTagsView to display errors for add, update, and remove operations.
Fixed order of operations when removing a tag and getting user confirmation.
Added unit test for DeviceTagCollection.updateTag()
Promoted DeviceModel.getAferoClient() to be public.